+++ src/sysReg.c
@@ -629,7 +629,7 @@ void Hardclock_InterruptHandler ( void )
         Log_Printf(LOG_DEBUG, "[INT] throwing hardclock %lld", host_time_us());
         set_interrupt(INT_TIMER,SET_INT);
         uint64_t now = host_time_us();
-        host_hardclock(latch_hardclock, now - hardClockLastLatch);
+        host_hardclock(latch_hardclock, (int)(now - hardClockLastLatch));
         hardClockLastLatch = now;
         CycInt_AddRelativeInterruptUs(latch_hardclock, 0, INTERRUPT_HARDCLOCK);
     }
